requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
Django==4.2.16
django-allauth==65.2.0
django-cloudinary-storage==0.3.0
pillow==11.0.0
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.2
